# docs/swagger.yaml
basePath: /
...
host: localhost:9999
info:
  contact: {}
  description: Authentication Service
  title: Autentico OIDC
  version: "1.0"
paths:
...
  /oauth2/passkey/login/begin:
    get:
      consumes:
      - application/json
      description: Initiates a WebAuthn authentication ceremony. The user must already
        have a registered passkey. Returns the options for the navigator.credentials.get
        call.
      parameters:
      - description: User's username
        in: query
        name: username
        required: true
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      - description: OAuth2 state
        in: query
        name: state
        type: string
      - description: OAuth2 client ID
        in: query
        name: client_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: WebAuthn assertion options
          schema:
            additionalProperties: true
            type: object
      summary: Begin passkey login
      tags:
      - passkey
  /oauth2/passkey/login/finish:
    post:
      consumes:
      - application/json
      description: Processes the WebAuthn assertion from the client and issues an
        authorization code.
      parameters:
      - description: Challenge ID from BeginLogin
        in: query
        name: challenge_id
        required: true
        type: string
      - description: WebAuthn assertion response
        in: body
        name: assertion
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Redirect URL
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Complete passkey login
      tags:
      - passkey
  /oauth2/passkey/register/begin:
    get:
      description: Creates a user account (if not already present) and initiates a
        WebAuthn registration ceremony.
      parameters:
      - description: Desired username
        in: query
        name: username
        required: true
        type: string
      - description: Email address
        in: query
        name: email
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      - description: OAuth2 state
        in: query
        name: state
        type: string
      - description: OAuth2 client ID
        in: query
        name: client_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: WebAuthn registration options
          schema:
            additionalProperties: true
            type: object
      summary: Begin passkey registration
      tags:
      - passkey
  /oauth2/passkey/register/finish:
    post:
      consumes:
      - application/json
      description: Processes the WebAuthn attestation from the client and registers
        the passkey.
      parameters:
      - description: Challenge ID from BeginRegistration
        in: query
        name: challenge_id
        required: true
        type: string
      - description: WebAuthn attestation response
        in: body
        name: attestation
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Redirect URL
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Complete passkey registration
      tags:
      - passkey
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"